+++ examples/analog-stopwatch-with-spi-ssd1306.rs
@@ -16,7 +16,6 @@ use crate::hal::{
     prelude::*,
     rcc::{Clocks, Rcc},
     spi::Spi,
-    stm32,
     timer::{CountDownTimer, Event, Timer},
 };
 
@@ -47,7 +46,7 @@ use ssd1306::{prelude::*, Builder};
 // Set up global state. It's all mutexed up for concurrency safety.
 static ELAPSED_MS: Mutex<Cell<u32>> = Mutex::new(Cell::new(0u32));
 static ELAPSED_RESET_MS: Mutex<Cell<u32>> = Mutex::new(Cell::new(0u32));
-static TIMER_TIM2: Mutex<RefCell<Option<CountDownTimer<stm32::TIM2>>>> =
+static TIMER_TIM2: Mutex<RefCell<Option<CountDownTimer<pac::TIM2>>>> =
     Mutex::new(RefCell::new(None));
 static STATE: Mutex<Cell<StopwatchState>> = Mutex::new(Cell::new(StopwatchState::Ready));
 static BUTTON: Mutex<RefCell<Option<PA0<Input<PullDown>>>>> = Mutex::new(RefCell::new(None));
@@ -139,10 +138,10 @@ fn main() -> ! {
     });
 
     // Enable interrupts
-    stm32::NVIC::unpend(hal::stm32::Interrupt::TIM2);
-    stm32::NVIC::unpend(hal::stm32::Interrupt::EXTI0);
+    pac::NVIC::unpend(hal::pac::Interrupt::TIM2);
+    pac::NVIC::unpend(hal::pac::Interrupt::EXTI0);
     unsafe {
-        stm32::NVIC::unmask(hal::stm32::Interrupt::EXTI0);
+        pac::NVIC::unmask(hal::pac::Interrupt::EXTI0);
     };
 
     let mut state_led = false;
@@ -169,11 +168,10 @@ fn main() -> ! {
                 led4.set_low();
             }
             StopwatchState::Running => {
+                led4.set_low();
                 if state_led {
-                    led4.set_low();
                     led3.set_high();
                 } else {
-                    led4.set_low();
                     led3.set_low();
                 }
             }
@@ -209,13 +207,12 @@ fn main() -> ! {
 }
 
 fn setup_clocks(rcc: Rcc) -> Clocks {
-    return rcc
-        .cfgr
+    rcc.cfgr
         .hclk(180.mhz())
         .sysclk(180.mhz())
         .pclk1(45.mhz())
         .pclk2(90.mhz())
-        .freeze();
+        .freeze()
 }
 
 #[interrupt]
@@ -292,32 +289,32 @@ fn TIM2() {
     });
 }
 
-fn stopwatch_start<'cs>(cs: &'cs CriticalSection) {
+fn stopwatch_start(cs: &CriticalSection) {
     ELAPSED_MS.borrow(cs).replace(0);
     unsafe {
-        stm32::NVIC::unmask(hal::stm32::Interrupt::TIM2);
+        pac::NVIC::unmask(hal::pac::Interrupt::TIM2);
     }
 }
 
-fn stopwatch_continue<'cs>(_cs: &'cs CriticalSection) {
+fn stopwatch_continue(_cs: &CriticalSection) {
     unsafe {
-        stm32::NVIC::unmask(hal::stm32::Interrupt::TIM2);
+        pac::NVIC::unmask(hal::pac::Interrupt::TIM2);
     }
 }
 
-fn stopwatch_stop<'cs>(_cs: &'cs CriticalSection) {
-    stm32::NVIC::mask(hal::stm32::Interrupt::TIM2);
+fn stopwatch_stop(_cs: &CriticalSection) {
+    pac::NVIC::mask(hal::pac::Interrupt::TIM2);
 }
 
-fn stopwatch_reset_start<'cs>(cs: &'cs CriticalSection) {
+fn stopwatch_reset_start(cs: &CriticalSection) {
     ELAPSED_RESET_MS.borrow(cs).replace(0);
     unsafe {
-        stm32::NVIC::unmask(hal::stm32::Interrupt::TIM2);
+        pac::NVIC::unmask(hal::pac::Interrupt::TIM2);
     }
 }
 
-fn stopwatch_reset_stop<'cs>(_cs: &'cs CriticalSection) {
-    stm32::NVIC::mask(hal::stm32::Interrupt::TIM2);
+fn stopwatch_reset_stop(_cs: &CriticalSection) {
+    pac::NVIC::mask(hal::pac::Interrupt::TIM2);
 }
 
 // Formatting requires the arrayvec crate
@@ -333,15 +330,15 @@ fn format_elapsed(buf: &mut ArrayString<[u8; 10]>, elapsed: u32) {
 }
 
 fn elapsed_to_ms(elapsed: u32) -> u32 {
-    return elapsed % 1000;
+    elapsed % 1000
 }
 
 fn elapsed_to_s(elapsed: u32) -> u32 {
-    return (elapsed - elapsed_to_ms(elapsed)) % 60000 / 1000;
+    (elapsed - elapsed_to_ms(elapsed)) % 60000 / 1000
 }
 
 fn elapsed_to_m(elapsed: u32) -> u32 {
-    return elapsed / 60000;
+    elapsed / 60000
 }
 
 /// Convert a polar coordinate (angle/distance) into an (X, Y) coordinate centered around `CENTER`