+++ examples/delay-syst-blinky.rs
@@ -1,22 +1,22 @@
 //! Demonstrate the use of a blocking `Delay` using the SYST (sysclock) timer.
 
 #![deny(unsafe_code)]
+#![allow(clippy::empty_loop)]
 #![no_main]
 #![no_std]
 
 // Halt on panic
 use panic_halt as _; // panic handler
 
-use cortex_m;
 use cortex_m_rt::entry;
 use stm32f4xx_hal as hal;
 
-use crate::hal::{prelude::*, stm32};
+use crate::hal::{pac, prelude::*};
 
 #[entry]
 fn main() -> ! {
     if let (Some(dp), Some(cp)) = (
-        stm32::Peripherals::take(),
+        pac::Peripherals::take(),
         cortex_m::peripheral::Peripherals::take(),
     ) {
         // Set up the LED. On the Nucleo-446RE it's connected to pin PA5.