+++ examples/pwm-input.rs
@@ -5,13 +5,12 @@
 // Halt on panic
 use panic_halt as _;
 
-use cortex_m;
 use cortex_m_rt::entry;
-use stm32f4xx_hal::{prelude::*, stm32, timer::Timer};
+use stm32f4xx_hal::{pac, prelude::*, timer::Timer};
 
 #[entry]
 fn main() -> ! {
-    if let Some(dp) = stm32::Peripherals::take() {
+    if let Some(dp) = pac::Peripherals::take() {
         // Set up the system clock.
         let rcc = dp.RCC.constrain();
         let clocks = rcc.cfgr.freeze();