+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Simple docs in gpio. `into_<output>_in_state`, `with_<output>_in_state`
+- Weaker constrains for examples.
+- Deprecate `stm32` alias.
 - Temporary change pin mode
 - More badges in README
 - `RccBus` & `GetBusFreq` traits. `AHBx`, `APBx` structures.
@@ -51,6 +54,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- [breaking-change] `into_<output>` fns set pin in `Low` state by default
 - Use manual impls for blocking spi instead of `Default`.
 - Split `Stream` trait on `Stream` and `StreamISR`.
   Use const generics for `Stream` and `Channel`.